+++ js/modules/geom/bundle.mjs
@@ -1,0 +1,5 @@
+// Only build function to create three.mjs model
+
+export { version, parse } from '../core.mjs';
+export { openFile } from '../io.mjs';
+export { build } from './TGeoPainter.mjs';